Initial script work
We're getting there


Further work


Nearly done with the cmake refactor


Finished refactor, need to finish precompiled headers


Updated luaJit and VkHLF, removed vkel, made changes to component factories


Precompiled files aren't showing up in the right filter and it's sad.


Finish implementation of precompiled headers, just needs real additions.

Also deleted some old files, changed some files to hpps that were h
Public headers weren't being added to the public precompiled header list


Removed some unneeded files.


Disabled multi-pass rendering as it's broken according to Validation Layers

Also made some changes to appease the ninja generator. Still doesn't work due to the precompiled source file not having a full path. Likely could fix if the generated file were put in the source tree rather than build tree.
Added fmt and fixed a missing library issue with luajit.


We are mostly running again.


Made MaterialViewer look like default levels, fixed Qt tab issue.


Worked on some logging system stuff, UTF8 stuff, and undefined behavior stuff.


Started thinking out a string view class


Initial ninja support.


Initial subpass


Fixed some things, editor running again.


A nice meme

Added Publish build config so we can change the config dir accordingly

A nice cmake meme

+++ Source/YTEditor/ComponentBrowser/PropertyWidget.hpp
@@ -367,7 +367,7 @@ namespace YTEditor
   template<>
   inline void PropertyWidget<glm::vec2>::SetValue(glm::vec2 aVal)
   {
-    for (int i = 0; i < mWidgets.size(); ++i)
+    for (size_t i = 0; i < mWidgets.size(); ++i)
     {
       LineEdit * widg = dynamic_cast<LineEdit*>(mWidgets[i]);
       char buff[20] = { '\0' };
@@ -379,7 +379,7 @@ namespace YTEditor
   template<>
   inline void PropertyWidget<glm::vec3>::SetValue(glm::vec3 aVal)
   {
-    for (int i = 0; i < mWidgets.size(); ++i)
+    for (size_t i = 0; i < mWidgets.size(); ++i)
     {
       LineEdit * widg = dynamic_cast<LineEdit*>(mWidgets[i]);
       char buff[20] = { '\0' };
@@ -391,7 +391,7 @@ namespace YTEditor
   template<>
   inline void PropertyWidget<glm::vec4>::SetValue(glm::vec4 aVal)
   {
-    for (int i = 0; i < mWidgets.size(); ++i)
+    for (size_t i = 0; i < mWidgets.size(); ++i)
     {
       LineEdit * widg = dynamic_cast<LineEdit*>(mWidgets[i]);
       char buff[20] = { '\0' };
